Move getIndex fn to alarm_logic to use across several components
