+++ src/alarm/alarm_logic.js
@@ -19,6 +19,19 @@ export const getTime = () => ({
   ms: new Date().getMilliseconds(),
 })
 
+export const getIndex = () => {
+  const time = getTime()
+  return Number(
+    time.y.toString() +
+    time.M.toString().padStart(2, "0") +
+    time.d.toString().padStart(2, "0") +
+    time.h.toString().padStart(2, "0") +
+    time.m.toString().padStart(2, "0") +
+    time.s.toString().padStart(2, "0") +
+    time.ms.toString().padStart(3, "0")
+  )
+}
+
 export const previousInterval = () => {
   let h = getTime().h
   let m = getTime().m
