+++ src/todo/todo_logic.js
@@ -1,19 +1,6 @@
 // src/todo/todo_logic.js
 
-import { getTime } from '../alarm/alarm_logic.js'
-
-const getIndex = () => {
-  const time = getTime()
-  return Number(
-    time.y.toString() +
-    time.M.toString().padStart(2, "0") +
-    time.d.toString().padStart(2, "0") +
-    time.h.toString().padStart(2, "0") +
-    time.m.toString().padStart(2, "0") +
-    time.s.toString().padStart(2, "0") +
-    time.ms.toString().padStart(3, "0")
-  )
-}
+import { getIndex } from '../alarm/alarm_logic.js'
 
 export const addTodo = (set, input) => set((state) => ({ todos: [...state.todos, { index: getIndex(), text: input }] }))
